btcjson/chainsvrresults: parse witness as string slice
This commit fixes a divergence between btcd and ltcd in
the expected witness return types when querying for
transactions over rpc. The fix returns the behavior to
match btcd, by treating witnesses as an array of strings
instead of a single string. This ensures that both
btcd and ltcd remain compatible from the client side.